requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
cachetools==5.3.2
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
cssselect==1.2.0
cssutils==2.9.0
Django==4.2.10
django-celery-beat==2.5.0
django-environ==0.11.2
django-timezone-field==6.1.0
djangorestframework==3.14.0
idna==3.6
kombu==5.3.5
lxml==5.1.0
premailer==3.10.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2024.1
redis==5.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
vine==5.1.0
wcwidth==0.2.13
yagmail==0.15.293

